Update DB_tables_constraints.sql
Reading up on some posts:
https://stackoverflow.com/questions/1481476/when-to-use-on-update-cascade
https://stackoverflow.com/questions/13444859/sql-on-delete-cascade-which-way-does-the-deletion-occur
https://stackoverflow.com/questions/6720050/foreign-key-constraints-when-to-use-on-update-and-on-delete
We had it kind of backwards, so I went in and fixed it. 


Users: It might be best if we just do not allow users to be deleted from the system. 
Employees: Taking out delete option. It should be default. Set it to inactive if they are no longer employed. 
Order: Again, instead of delete cascade, in the status we can notify that it is a cancelled order.